added sumNQTStringToNumber (util)
