+++ lib/packages/util/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## unreleased
+- added `sumNQTStringToNumber`
+
 ## 0.1.0-rc.3
 - Timestamp conversion functions
